+++ core-ng/src/main/java/core/framework/api/search/BulkIndexRequest.java
@@ -1,0 +1,11 @@
+package core.framework.api.search;
+
+import java.util.Map;
+
+/**
+ * @author neo
+ */
+public class BulkIndexRequest<T> {
+    public String index;
+    public Map<String, T> sources;
+}