+++ core-ng/src/main/java/core/framework/api/search/DeleteRequest.java
@@ -1,0 +1,9 @@
+package core.framework.api.search;
+
+/**
+ * @author neo
+ */
+public class DeleteRequest {
+    public String index;
+    public String id;
+}