+++ core-ng/src/main/java/core/framework/api/search/IndexRequest.java
@@ -1,0 +1,10 @@
+package core.framework.api.search;
+
+/**
+ * @author neo
+ */
+public class IndexRequest<T> {
+    public String index;
+    public String id;
+    public T source;
+}