+++ core-ng/src/main/java/core/framework/impl/log/DefaultLoggerFactory.java
@@ -24,17 +24,16 @@ public final class DefaultLoggerFactory implements ILoggerFactory {
     }
 
     private Logger createLogger(String name) {
-        LogLevel[] levels = logLevel(name);
-        return new LoggerImpl(name, logManager, levels[0], levels[1]);
+        return new LoggerImpl(name, logManager, traceLevel(name));
     }
 
-    private LogLevel[] logLevel(String name) {
+    private LogLevel traceLevel(String name) {
         if (name.startsWith("org.elasticsearch")
             || name.startsWith("org.mongodb")
             || name.startsWith("org.apache")
             || name.startsWith("org.xnio")) {
-            return new LogLevel[]{LogLevel.WARN, LogLevel.INFO};
+            return LogLevel.INFO;
         }
-        return new LogLevel[]{LogLevel.INFO, LogLevel.DEBUG};
+        return LogLevel.DEBUG;
     }
 }