+++ core-ng/src/main/java/core/framework/impl/log/LoggerImpl.java
@@ -28,15 +28,13 @@ final class LoggerImpl extends AbstractLogger {
 
     private final PrintStream output = System.out;
     private final LogManager logManager;
-    private final LogLevel logLevel;
     private final LogLevel traceLevel;
     private final String logger;
 
-    public LoggerImpl(String name, LogManager logManager, LogLevel logLevel, LogLevel traceLevel) {
+    public LoggerImpl(String name, LogManager logManager, LogLevel traceLevel) {
         super(name);
         this.logger = abbreviateLoggerName(name);
         this.logManager = logManager;
-        this.logLevel = logLevel;
         this.traceLevel = traceLevel;
     }
 
@@ -46,7 +44,7 @@ final class LoggerImpl extends AbstractLogger {
             LogEvent event = new LogEvent(logger, marker, level, message, arguments, exception);
             logManager.process(event);
 
-            if (level.value >= logLevel.value) {
+            if (level.value >= LogLevel.INFO.value) {
                 output.print(event.logMessage());
             }
         }