+++ core-ng/src/main/java/core/framework/impl/search/ElasticSearchTypeImpl.java
@@ -2,8 +2,14 @@ package core.framework.impl.search;
 
 import core.framework.api.log.ActionLogContext;
 import core.framework.api.log.Markers;
+import core.framework.api.search.BulkIndexRequest;
+import core.framework.api.search.DeleteRequest;
 import core.framework.api.search.ElasticSearchType;
+import core.framework.api.search.GetRequest;
+import core.framework.api.search.Index;
+import core.framework.api.search.IndexRequest;
 import core.framework.api.search.SearchException;
+import core.framework.api.search.SearchRequest;
 import core.framework.api.util.JSON;
 import core.framework.api.util.StopWatch;
 import core.framework.api.util.Strings;
@@ -12,11 +18,9 @@ import org.elasticsearch.action.bulk.BulkRequestBuilder;
 import org.elasticsearch.action.bulk.BulkResponse;
 import org.elasticsearch.action.delete.DeleteResponse;
 import org.elasticsearch.action.get.GetResponse;
-import org.elasticsearch.action.search.SearchRequest;
+import org.elasticsearch.action.search.SearchRequestBuilder;
 import org.elasticsearch.action.search.SearchResponse;
-import org.elasticsearch.action.update.UpdateRequest;
 import org.elasticsearch.client.Client;
-import org.elasticsearch.search.builder.SearchSourceBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -37,23 +41,69 @@ public final class ElasticSearchTypeImpl<T> implements ElasticSearchType<T> {
     private final long slowQueryThresholdInMs;
     private final Class<T> documentClass;
 
-    ElasticSearchTypeImpl(Client client, String index, String type, Class<T> documentClass, Duration slowQueryThreshold) {
+    ElasticSearchTypeImpl(Client client, Class<T> documentClass, Duration slowQueryThreshold) {
         this.client = client;
-        this.index = index;
-        this.type = type;
         this.documentClass = documentClass;
         this.slowQueryThresholdInMs = slowQueryThreshold.toMillis();
+        Index index = documentClass.getDeclaredAnnotation(Index.class);
+        this.index = index.index();
+        this.type = index.type();
         validator = new DocumentValidator<>(documentClass);
     }
 
     @Override
-    public void index(String id, T source) {
+    public SearchResponse search(SearchRequest request) {
         StopWatch watch = new StopWatch();
-        validator.validate(source);
+        long esTookTime = 0;
+        String index = request.index == null ? this.index : request.index;
         try {
-            String document = JSON.toJSON(source);
-            client.prepareIndex(index, type)
-                .setId(id)
+            SearchRequestBuilder builder = client.prepareSearch(index)
+                .setQuery(request.query);
+            request.aggregations.forEach(builder::addAggregation);
+            request.sorts.forEach(builder::addSort);
+            if (request.skip != null) builder.setFrom(request.skip);
+            if (request.limit != null) builder.setSize(request.limit);
+            logger.debug("search, index={}, type={}, request={}", index, type, builder);
+            SearchResponse response = builder.get();
+            esTookTime = response.getTookInMillis();
+            if (response.getFailedShards() > 0) logger.warn("some shard failed, response={}", response);
+            return response;
+        } catch (ElasticsearchException e) {
+            throw new SearchException(e);   // due to elastic search uses async executor to run, we have to wrap the exception to retain the original place caused the exception
+        } finally {
+            long elapsedTime = watch.elapsedTime();
+            ActionLogContext.track("elasticsearch", elapsedTime);
+            logger.debug("search, esTookTime={}, elapsedTime={}", esTookTime, elapsedTime);
+            checkSlowQuery(elapsedTime);
+        }
+    }
+
+    @Override
+    public Optional<T> get(GetRequest request) {
+        StopWatch watch = new StopWatch();
+        String index = request.index == null ? this.index : request.index;
+        try {
+            GetResponse response = client.prepareGet(index, type, request.id).get();
+            if (!response.isExists()) return Optional.empty();
+            return Optional.of(JSON.fromJSON(documentClass, response.getSourceAsString()));
+        } catch (ElasticsearchException e) {
+            throw new SearchException(e);   // due to elastic search uses async executor to run, we have to wrap the exception to retain the original place caused the exception
+        } finally {
+            long elapsedTime = watch.elapsedTime();
+            ActionLogContext.track("elasticsearch", elapsedTime);
+            logger.debug("get, index={}, type={}, id={}, elapsedTime={}", index, type, request.id, elapsedTime);
+            checkSlowQuery(elapsedTime);
+        }
+    }
+
+    @Override
+    public void index(IndexRequest<T> request) {
+        StopWatch watch = new StopWatch();
+        String index = request.index == null ? this.index : request.index;
+        validator.validate(request.source);
+        try {
+            String document = JSON.toJSON(request.source);
+            client.prepareIndex(index, type, request.id)
                 .setSource(Strings.bytes(document))
                 .get();
         } catch (ElasticsearchException e) {
@@ -61,24 +111,24 @@ public final class ElasticSearchTypeImpl<T> implements ElasticSearchType<T> {
         } finally {
             long elapsedTime = watch.elapsedTime();
             ActionLogContext.track("elasticsearch", elapsedTime);
-            logger.debug("index, index={}, type={}, id={}, elapsedTime={}", index, type, id, elapsedTime);
+            logger.debug("index, index={}, type={}, id={}, elapsedTime={}", index, type, request.id, elapsedTime);
             checkSlowQuery(elapsedTime);
         }
     }
 
     @Override
-    public void bulkIndex(Map<String, T> sources) {
+    public void bulkIndex(BulkIndexRequest<T> request) {
         StopWatch watch = new StopWatch();
+        String index = request.index == null ? this.index : request.index;
         long esTookTime = 0;
         try {
             BulkRequestBuilder builder = client.prepareBulk();
-            for (Map.Entry<String, T> entry : sources.entrySet()) {
+            for (Map.Entry<String, T> entry : request.sources.entrySet()) {
                 String id = entry.getKey();
                 T source = entry.getValue();
                 validator.validate(source);
                 String document = JSON.toJSON(source);
-                builder.add(client.prepareIndex(index, type)
-                    .setId(id)
+                builder.add(client.prepareIndex(index, type, id)
                     .setSource(Strings.bytes(document)));
             }
             BulkResponse response = builder.get();
@@ -89,81 +139,24 @@ public final class ElasticSearchTypeImpl<T> implements ElasticSearchType<T> {
         } finally {
             long elapsedTime = watch.elapsedTime();
             ActionLogContext.track("elasticsearch", elapsedTime);
-            logger.debug("bulkIndex, index={}, type={}, size={}, esTookTime={}, elapsedTime={}", index, type, sources.size(), esTookTime, elapsedTime);
+            logger.debug("bulkIndex, index={}, type={}, size={}, esTookTime={}, elapsedTime={}", index, type, request.sources.size(), esTookTime, elapsedTime);
             checkSlowQuery(elapsedTime);
         }
     }
 
     @Override
-    public Optional<T> get(String id) {
+    public boolean delete(DeleteRequest request) {
         StopWatch watch = new StopWatch();
+        String index = request.index == null ? this.index : request.index;
         try {
-            GetResponse response = client.prepareGet(index, type, id).get();
-            if (!response.isExists()) return Optional.empty();
-            return Optional.of(JSON.fromJSON(documentClass, response.getSourceAsString()));
-        } catch (ElasticsearchException e) {
-            throw new SearchException(e);   // due to elastic search uses async executor to run, we have to wrap the exception to retain the original place caused the exception
-        } finally {
-            long elapsedTime = watch.elapsedTime();
-            ActionLogContext.track("elasticsearch", elapsedTime);
-            logger.debug("get, index={}, type={}, id={}, elapsedTime={}", index, type, id, elapsedTime);
-            checkSlowQuery(elapsedTime);
-        }
-    }
-
-    @Override
-    public void update(String id, UpdateRequest request) {
-        StopWatch watch = new StopWatch();
-        try {
-            request.index(index)
-                .type(type)
-                .id(id);
-            client.update(request).actionGet();
-        } catch (ElasticsearchException e) {
-            throw new SearchException(e);   // due to elastic search uses async executor to run, we have to wrap the exception to retain the original place caused the exception
-        } finally {
-            long elapsedTime = watch.elapsedTime();
-            ActionLogContext.track("elasticsearch", elapsedTime);
-            logger.debug("update, index={}, type={}, id={}, elapsedTime={}", index, type, id, elapsedTime);
-            checkSlowQuery(elapsedTime);
-        }
-    }
-
-    @Override
-    public boolean delete(String id) {
-        StopWatch watch = new StopWatch();
-        try {
-            DeleteResponse response = client.prepareDelete(index, type, id).get();
+            DeleteResponse response = client.prepareDelete(index, type, request.id).get();
             return response.isFound();
         } catch (ElasticsearchException e) {
             throw new SearchException(e);   // due to elastic search uses async executor to run, we have to wrap the exception to retain the original place caused the exception
         } finally {
             long elapsedTime = watch.elapsedTime();
             ActionLogContext.track("elasticsearch", elapsedTime);
-            logger.debug("delete, index={}, type={}, id={}, elapsedTime={}", index, type, id, elapsedTime);
-            checkSlowQuery(elapsedTime);
-        }
-    }
-
-    @Override
-    public SearchResponse search(SearchSourceBuilder source) {
-        StopWatch watch = new StopWatch();
-        long esTookTime = 0;
-        try {
-            logger.debug("search, index={}, type={}, source={}", index, type, source);
-            SearchRequest request = new SearchRequest(index)
-                .source(source)
-                .types(type);
-            SearchResponse response = client.search(request).actionGet();
-            esTookTime = response.getTookInMillis();
-            if (response.getFailedShards() > 0) logger.warn("some shard failed, response={}", response);
-            return response;
-        } catch (ElasticsearchException e) {
-            throw new SearchException(e);   // due to elastic search uses async executor to run, we have to wrap the exception to retain the original place caused the exception
-        } finally {
-            long elapsedTime = watch.elapsedTime();
-            ActionLogContext.track("elasticsearch", elapsedTime);
-            logger.debug("search, esTookTime={}, elapsedTime={}", esTookTime, elapsedTime);
+            logger.debug("delete, index={}, type={}, id={}, elapsedTime={}", index, type, request.id, elapsedTime);
             checkSlowQuery(elapsedTime);
         }
     }