+++ log-processor/src/test/java/core/log/queue/ActionLogMessagesHandlerTest.java
@@ -1,17 +1,21 @@
 package core.log.queue;
 
 import core.framework.api.search.ElasticSearchType;
+import core.framework.api.search.GetRequest;
 import core.framework.api.util.Lists;
 import core.framework.impl.log.queue.ActionLogMessage;
 import core.framework.impl.log.queue.ActionLogMessages;
 import core.log.IntegrationTest;
 import core.log.domain.ActionLogDocument;
 import core.log.domain.TraceLogDocument;
-import org.junit.Assert;
 import org.junit.Test;
 
 import javax.inject.Inject;
 import java.time.Instant;
+import java.time.LocalDate;
+import java.time.Month;
+
+import static org.junit.Assert.assertEquals;
 
 /**
  * @author neo
@@ -40,13 +44,25 @@ public class ActionLogMessagesHandlerTest extends IntegrationTest {
 
         ActionLogMessages messages = new ActionLogMessages();
         messages.logs = Lists.newArrayList(message1, message2);
-        handler.handle(messages);
+        LocalDate now = LocalDate.of(2016, Month.JANUARY, 15);
+        handler.handle(messages, now);
 
-        ActionLogDocument action = actionType.get(message1.id).get();
-        Assert.assertEquals(message1.result, action.result);
+        GetRequest request = new GetRequest();
+        request.index = handler.index("action", now);
+        request.id = message1.id;
+        ActionLogDocument action = actionType.get(request).get();
+        assertEquals(message1.result, action.result);
 
-        TraceLogDocument trace = traceType.get(message2.id).get();
-        Assert.assertEquals(message2.id, trace.id);
-        Assert.assertEquals(message2.traceLog, trace.content);
+        request = new GetRequest();
+        request.index = handler.index("trace", now);
+        request.id = message2.id;
+        TraceLogDocument trace = traceType.get(request).get();
+        assertEquals(message2.id, trace.id);
+        assertEquals(message2.traceLog, trace.content);
+    }
+
+    @Test
+    public void index() {
+        assertEquals("action-2016-01-15", handler.index("action", LocalDate.of(2016, Month.JANUARY, 15)));
     }
 }